# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSSD-APP API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://b-shop-api-5ntn.onrender.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/send-mail": {
            "post": {
                "tags": [
                    "Mailer"
                ],
                "summary": "Send an email",
                "description": "Send an email to the specified recipient.",
                "operationId": "sendMail",
                "requestBody": {
                    "description": "Email details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "to",
                                    "subject",
                                    "body"
                                ],
                                "properties": {
                                    "to": {
                                        "type": "string",
                                        "example": "recipient@example.com"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Subject of the email"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Body of the email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email sent successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to send email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get list of notifications",
                "description": "Returns list of notifications",
                "operationId": "0f031f66cd8c52bf73e89fb2d050ad83",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notification"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a new notification",
                "description": "Create a new notification",
                "operationId": "258ba7c348fd4283c6b3802a69fe2fb0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Notification"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get specified notification",
                "description": "Returns specified notification",
                "operationId": "0c8e017ab6d765f82f97368b80f52b8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Update specified notification",
                "description": "Update specified notification",
                "operationId": "38c5f13acd3d0b84545d309539e6b972",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Notification"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete specified notification",
                "description": "Delete specified notification",
                "operationId": "ccdcc730b2088097dd94128f24960dbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Notification": {
                "title": "Notification",
                "description": "Notification model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the notification",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID associated with the notification",
                        "type": "integer",
                        "example": "1"
                    },
                    "message": {
                        "description": "Content of the notification",
                        "type": "string",
                        "example": "You have a new message."
                    },
                    "notification_type": {
                        "description": "Type of the notification",
                        "type": "string",
                        "example": "email"
                    },
                    "created_at": {
                        "description": "Date and time when the notification was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 08:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the notification was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 08:30:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the user",
                        "type": "integer",
                        "example": "1"
                    },
                    "full_name": {
                        "description": "The full name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "username": {
                        "description": "The username of the user",
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "description": "The password of the user (hashed)",
                        "type": "string",
                        "format": "password",
                        "example": "$2y$10$N0wbN10w.0.yBhEWwH2fveCAeAxFxuIQJ4O7QHBTQm7v9eN6LpfSi"
                    },
                    "phone_number": {
                        "description": "The phone number of the user",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email_verified_at": {
                        "description": "The timestamp of email verification",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    },
                    "created_at": {
                        "description": "The timestamp of when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    },
                    "updated_at": {
                        "description": "The timestamp of when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}